# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Intell Meet API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/users/{userId}/role:
    put:
      operationId: adminUpdateUserRole
      tags: [analytics]
      summary: Update user platform role (Admin only)
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUpdateUserRoleInput"
      responses:
          "200":
            description: Role updated
            content:
              application/json:
                schema:
                  $ref: "#/components/schemas/User"
  /admin/users/{userId}:
    delete:
      operationId: adminDeleteUser
      tags: [analytics]
      summary: Delete a user account (Admin only)
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                required: [message]
...
components:
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [Admin, Manager, Member]
        phoneNumber:
          type: ["string", "null"]
        jobTitle:
          type: ["string", "null"]
        department:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        timezone:
          type: ["string", "null"]
        avatar:
          type: ["string", "null"]
        authProvider:
          type: string
          enum: [local, google]
        googleId:
          type: ["string", "null"]
        profilePicture:
          type: ["string", "null"]
        emailVerified:
          type: boolean
        hasPassword:
          type: boolean
        notificationSettings:
          type: object
          properties:
            email:
              type: boolean
            push:
              type: boolean
            sms:
              type: boolean
          required: [email, push, sms]
        createdAt:
          type: string
      required: [id, name, email, role, createdAt, authProvider, emailVerified, hasPassword]
...
    AdminUpdateUserRoleInput:
      type: object
      properties:
        role:
          type: string
          enum: [Admin, Manager, Member]
      required: [role]
...